+++ tensorflow_datasets/image/image_folder.py
@@ -74,7 +74,8 @@ class ImageLabelFolder(tfds.core.GeneratorBasedBuilder):
 
   """
 
-  VERSION = tfds.core.Version("1.0.0")
+  VERSION = tfds.core.Version("1.0.0",
+                              experiments={tfds.core.Experiment.S3: False})
 
   # TODO(epot): Image shape should be automatically deduced