+++ tensorflow_datasets/public_api.py
@@ -21,6 +21,7 @@ tf_compat.ensure_tf_install()
 
 from tensorflow_datasets import core
 from tensorflow_datasets.core import download
+from tensorflow_datasets.core import decode
 from tensorflow_datasets.core import features
 from tensorflow_datasets.core import file_format_adapter as file_adapter
 from tensorflow_datasets.core import units
@@ -39,6 +40,7 @@ from tensorflow_datasets.version import __version__
 __all__ = [
     "core",
     "as_numpy",
+    "decode",
     "download",
     "features",
     "file_adapter",
@@ -52,6 +54,7 @@ __all__ = [
     "testing",
     "disable_progress_bar",
     "is_dataset_on_gcs",
+    "__version__",
 ]