+++ tensorflow_datasets/structured/higgs.py
@@ -61,7 +61,8 @@ _URL = 'https://archive.ics.uci.edu/ml/machine-learning-databases/00280/HIGGS.cs
 class Higgs(tfds.core.GeneratorBasedBuilder):
   """HIGGS Data Set."""
 
-  VERSION = tfds.core.Version('1.0.0')
+  VERSION = tfds.core.Version('1.0.0',
+                              experiments={tfds.core.Experiment.S3: False})
 
   def _info(self):
     return tfds.core.DatasetInfo(