+++ docs/release_notes.md
@@ -6,8 +6,17 @@
 
 *   Image:
     [downsampled_imagenet](https://github.com/tensorflow/datasets/tree/master/docs/datasets.md#downsampled_imagenet)
+*   Image:
+    [patch_camelyon](https://github.com/tensorflow/datasets/tree/master/docs/datasets.md#patch_camelyon)
+*   Image:
+    [binarized_mnist](https://github.com/tensorflow/datasets/tree/master/docs/datasets.md#binarized_mnist)
 
 ### Features
 
+*   Add `in_memory` option to cache small dataset in RAM.
+*   Better sharding, shuffling and sub-split
 *   It is now possible to add arbitrary metadata to `tfds.core.DatasetInfo`
     which will be stored/restored with the dataset. See `tfds.core.Metadata`.
+*   Better proxy support, possibility to add certificate
+*   Add `decoders` kwargs to override the default feature decoding
+    ([guide](https://github.com/tensorflow/datasets/tree/master/docs/decode.md)).